- Fixing warnings in regex
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@15387 c90b9560-bf6c-de11-be94-00142212c4b1

+++ regex/engine.c
@@ -71,7 +71,7 @@ regmatch_t pmatch[];
 int eflags;
 {
 	register char *endp;
-	register int i;
+	register size_t i;
 	struct match mv;
 	register struct match *m = &mv;
 	register char *dp;
